EZP-25101: Refactored ContentField validator
ContentFieldValidator is removed.
FieldValueValidator is not abstract anymore, and can be used to validate
any FieldValue. DefaultFieldValueValidator extends FieldValueValidator.
