+++ lib/Validator/Constraints/FieldValueValidator.php
@@ -11,12 +11,13 @@
 namespace EzSystems\RepositoryForms\Validator\Constraints;
 
 use eZ\Publish\API\Repository\Values\ValueObject;
+use EzSystems\RepositoryForms\Data\Content\FieldData;
 use Symfony\Component\Validator\Constraint;
 
 /**
  * Base class for field value validators.
  */
-abstract class FieldValueValidator extends FieldTypeValidator
+class FieldValueValidator extends FieldTypeValidator
 {
     public function validate($value, Constraint $constraint)
     {
@@ -36,34 +37,48 @@ abstract class FieldValueValidator extends FieldTypeValidator
     }
 
     /**
-     * Returns the field value to validate, or null if there is nothing to validate (e.g. empty default value).
+     * Returns the field value to validate.
      *
-     * @param ValueObject $value ValueObject holding the field value to validate, e.g. FieldDefinitionData.
+     * @param FieldData|ValueObject $value FieldData ValueObject holding the field value to validate.
      *
      * @throws \InvalidArgumentException If field value cannot be retrieved.
      *
-     * @return \eZ\Publish\SPI\FieldType\Value|null
+     * @return \eZ\Publish\SPI\FieldType\Value
      */
-    abstract protected function getFieldValue(ValueObject $value);
+    protected function getFieldValue(ValueObject $value)
+    {
+        return $value->value;
+    }
 
     /**
      * Returns the field definition $value refers to.
      * FieldDefinition object is needed to validate field value against field settings.
      *
-     * @param ValueObject $value ValueObject holding the field value to validate, e.g. FieldDefinitionData.
+     * @param FieldData|ValueObject $value ValueObject holding the field value to validate, e.g. FieldDefinitionData.
      *
      * @throws \InvalidArgumentException If field definition cannot be retrieved.
      *
-     * @return \eZ\Publish\SPI\FieldType\Value
+     * @return \eZ\Publish\API\Repository\Values\ContentType\FieldDefinition
      */
-    abstract protected function getFieldDefinition(ValueObject $value);
+    protected function getFieldDefinition(ValueObject $value)
+    {
+        return $value->fieldDefinition;
+    }
 
     /**
      * Returns the fieldTypeIdentifier for the field value to validate.
      *
-     * @param ValueObject $value ValueObject holding the field value to validate, e.g. FieldDefinitionData.
+     * @param FieldData|ValueObject $value FieldData ValueObject holding the field value to validate.
      *
      * @return string
      */
-    abstract protected function getFieldTypeIdentifier(ValueObject $value);
+    protected function getFieldTypeIdentifier(ValueObject $value)
+    {
+        return $value->fieldDefinition->fieldTypeIdentifier;
+    }
+
+    protected function generatePropertyPath($errorIndex, $errorTarget)
+    {
+        return 'value';
+    }
 }
